+++ web/src/html_cd_control.c
@@ -6,7 +6,7 @@
 
 static void create_html(char *msg) {
   cdrom_drive *d;
-  unsigned long msg_end;
+  size_t msg_end;
   strcpy(msg, "<!DOCTYPE html>");
   strcat(msg, "<html lang=en>");
   strcat(msg, "<head>");
@@ -40,7 +40,7 @@ static void create_html(char *msg) {
 }
 
 static void create_header(char *hdr, unsigned long msg_len) {
-  unsigned long hdr_end;
+  size_t hdr_end;
   strcpy(hdr, "HTTP/1.1 200 OK\r\n");
   strcat(hdr, "Content-Type: text/html; charset=utf-8\r\n");
   strcat(hdr, "Cache-control: no-cache\r\n");
